chore: move isSystemDirectory and isWritable to fileUtils
No behavior change. Drops the mcp -> backend coupling that was only
there for the path heuristic.
