+++ packages/playwright-core/src/tools/backend/context.ts
@@ -22,7 +22,7 @@ import debug from 'debug';
 import { escapeWithQuotes } from '@isomorphic/stringUtils';
 import { disposeAll } from '@isomorphic/disposable';
 import { eventsHelper } from '@utils/eventsHelper';
-import { isPathInside } from '@utils/fileUtils';
+import { isPathInside, isSystemDirectory, isWritable } from '@utils/fileUtils';
 import { playwright } from '../../inprocess';
 
 import { Tab } from './tab';
@@ -385,32 +385,13 @@ export async function workspaceFile(options: ContextOptions, fileName: string, p
   return resolvedName;
 }
 
-export function isUnsuitableForOutput(dir: string): boolean {
-  const resolved = path.resolve(dir);
-  if (process.platform === 'win32') {
-    const systemRoot = path.resolve(process.env.SystemRoot || 'C:\\Windows');
-    return isPathInside(systemRoot.toLowerCase(), resolved.toLowerCase());
-  }
-  return resolved === '/';
-}
-
-function isWritableDir(dir: string): boolean {
-  try {
-    fs.accessSync(dir, fs.constants.W_OK);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
 export function outputDir(options: ContextOptions): string {
   if (options.config.outputDir)
     return path.resolve(options.config.outputDir);
   const baseName = options.config.skillMode ? '.playwright-cli' : '.playwright-mcp';
-  const cwd = path.resolve(options.cwd);
-  if (isUnsuitableForOutput(cwd) || !isWritableDir(cwd))
+  if (isSystemDirectory(options.cwd) || !isWritable(options.cwd))
     return path.join(os.tmpdir(), baseName);
-  return path.join(cwd, baseName);
+  return path.join(options.cwd, baseName);
 }
 
 export async function outputFile(options: ContextOptions, fileName: string, flags: { origin: 'code' | 'llm' }): Promise<string> {
