+++ tests/mcp/config-resolve.spec.ts
@@ -23,7 +23,7 @@ import { test, expect } from './fixtures';
 import { tools } from '../../packages/playwright-core/lib/coreBundle';
 import type { Config } from '../../packages/playwright-core/src/tools/mcp/config.d';
 
-const { resolveCLIConfigForCLI, resolveCLIConfigForMCP, isUnsuitableForOutput, outputDir } = tools;
+const { resolveCLIConfigForCLI, resolveCLIConfigForMCP, isSystemDirectory, outputDir } = tools;
 
 // Empty env to isolate tests from the host environment.
 const emptyEnv = {};
@@ -310,10 +310,10 @@ test.describe('outputDir', () => {
     }
   });
 
-  test('isUnsuitableForOutput detects /', () => {
-    expect(isUnsuitableForOutput('/')).toBe(true);
-    expect(isUnsuitableForOutput('/tmp')).toBe(false);
-    expect(isUnsuitableForOutput(os.homedir())).toBe(false);
+  test('isSystemDirectory detects /', () => {
+    expect(isSystemDirectory('/')).toBe(true);
+    expect(isSystemDirectory('/tmp')).toBe(false);
+    expect(isSystemDirectory(os.homedir())).toBe(false);
   });
 
   test('resolveCLIConfigForMCP throws when --output-dir is /', async () => {
